+++ test/util/mosquitto.py
@@ -3,7 +3,7 @@
 # (c) 2020 Andreas Motl <andreas@hiveeyes.org>
 # License: GNU General Public License, Version 3
 import os
-
+import socket
 import pytest
 from pytest_docker_fixtures.containers._base import BaseImage
 
@@ -44,8 +44,32 @@ class Mosquitto(BaseImage):
 mosquitto_image = Mosquitto()
 
 
+def is_port_reachable(host, port):
+    """
+    Test if a host is up.
+    https://github.com/lovelysystems/lovely.testlayers/blob/0.7.0/src/lovely/testlayers/util.py#L6-L13
+    """
+    s = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
+    ex = s.connect_ex((host, port))
+    if ex == 0:
+        s.close()
+        return True
+    return False
+
+
+def is_mosquitto_running():
+    return is_port_reachable('localhost', 1883)
+
+
 @pytest.fixture(scope='session')
 def mosquitto():
+
+    # Gracefully skip spinning up the Docker container if Mosquitto is already running.
+    if is_mosquitto_running():
+        yield
+        return
+
+    # Spin up Mosquitto container.
     if os.environ.get('MOSQUITTO'):
         yield os.environ['MOSQUITTO'].split(':')
     else: